+++ src/buffer/mapshaper-polyline-buffer.js
@@ -1,9 +1,10 @@
 import { getPolylineBufferMaker } from '../buffer/mapshaper-path-buffer';
 import { getPolylineBufferMaker2 } from '../buffer/mapshaper-path-buffer2';
 import { getBearingFunction } from '../geom/mapshaper-geodesic';
-import { getGeodeticSegmentFunction } from '../geom/mapshaper-geodesic';
+import { getFastGeodeticSegmentFunction } from '../geom/mapshaper-geodesic';
 import { getBufferToleranceFunction, getBufferDistanceFunction, dissolveBufferDataset } from '../buffer/mapshaper-buffer-common';
 import { importGeoJSON } from '../geojson/geojson-import';
+import { getDatasetCRS } from '../crs/mapshaper-projections';
 
 export function makePolylineBuffer(lyr, dataset, opts) {
   var geojson = makeShapeBufferGeoJSON(lyr, dataset, opts);
@@ -15,7 +16,7 @@ export function makePolylineBuffer(lyr, dataset, opts) {
 export function makeShapeBufferGeoJSON(lyr, dataset, opts) {
   var distanceFn = getBufferDistanceFunction(lyr, dataset, opts);
   var toleranceFn = getBufferToleranceFunction(dataset, opts);
-  var geod = getGeodeticSegmentFunction(dataset, false);
+  var geod = getFastGeodeticSegmentFunction(getDatasetCRS(dataset));
   var getBearing = getBearingFunction(dataset);
   var makerOpts = Object.assign({geometry_type: lyr.geometry_type}, opts);
   var factory = opts.v2 ? getPolylineBufferMaker2 : getPolylineBufferMaker;