+++ src/crs/mapshaper-proj-utils.js
@@ -1,0 +1,5 @@
+
+export function getSemiMinorAxis(P) {
+  return P.a * Math.sqrt(1 - (P.es || 0));
+}
+